:recycle: [Refactor] add infinite scroll on mover/my-quote page
